feat(migrations): create migration and seeding script
