+++ src/db/index.ts
@@ -2,4 +2,4 @@ import { Database } from "bun:sqlite"
 import { drizzle } from "drizzle-orm/bun-sqlite"
 
 const sqlite = new Database("sqlite.db")
-export const db = drizzle(sqlite)
+export const db = drizzle({ client: sqlite })
