+++ src/db/services/queue-number.service.ts
@@ -1,0 +1,11 @@
+import { db } from ".."
+import { courses } from "../models/courses.model"
+import { eq } from "drizzle-orm"
+
+export const queueNumberService = {
+  async findById() {
+    const result = await db.select().from(courses).where(eq(courses.id, 1))
+
+    console.log(result)
+  },
+}
