+++ src/db/utils/migrate.ts
@@ -1,0 +1,4 @@
+import { db } from ".."
+import { migrate } from "drizzle-orm/bun-sqlite/migrator"
+
+await migrate(db, { migrationsFolder: "./drizzle" })
