+++ src/db/utils/seed.ts
@@ -1,0 +1,19 @@
+import { db } from ".."
+import { courses } from "../models/courses.model"
+
+await db.insert(courses).values([
+  {
+    title: "The Matrix",
+    releaseYear: 1999,
+  },
+  {
+    title: "The Matrix Reloaded",
+    releaseYear: 2003,
+  },
+  {
+    title: "The Matrix Revolutions",
+    releaseYear: 2003,
+  },
+])
+
+console.log(`Seeding complete.`)
