+++ src/types/entities/course.ts
@@ -1,0 +1,5 @@
+export type Course = {
+  id: number
+  course_name: string
+  coordinator_availability: boolean
+}
