+++ src/types/entities/queue-number.ts
@@ -1,0 +1,6 @@
+export type QueueNumber = {
+  id: number
+  student_id: number
+  course_id: number
+  queue_number: number
+}
